Upload not working properly (#1)
Fixed the not properly working upload by fixing the API error within the
implementation. We falsely used the same API for groups and projects,
which is of course wrong.
Furthermore for uploading we should be only allowed to use the projects
endpoint. Besides those improvements we also added better logging and
extracted some logic.