Gradle Kotlin DSL Repository Resolution Fix
The Kotlin DSL had issues with fetching and uploading dependencies,
within #14 we provided a workaround. During the investigation we
realized that the whole repository resolution was buggy for Kotlin DSL.

With this changes we are adding tests for fetching dependencies,
as well as for uploading.
We achieved this by implementing the Action interface directly within the RepositoryActionHandler.
Furthermore we removed the deprecated `upload` method with the `owner`.

Closes: #14, #15