Issue/refactoring of tokens (#9)
* Refactoring Token Handling

The initial approach relied on subclassing for not really good reasons,
actually none. Instead of subclassing we changed the implementation to
Enums which makes the whole handling way easier.
There is still a lot of clean up needed. but as a first draft this looks
promising.

As more and more issues are coming in, we also restructured the tests
which should allow easier creating of functional tests.